requirements.txt
torch
torchaudio
transformers
numpy
tqdm
wandb
safetensors
audiomentations
Pillow
pyctcdecode
torchcodec
jiwer
librosa